Automate bootstrap with use of passfile

+++ bootstrap.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
 
+if [ ! -f "./passfile" ]; then
+  echo "Please create a file called ./passfile with a secure password"
+  exit 1
+fi
+
 mkdir -p certs crl newcerts private
 chmod 700 private
 touch index.txt
 echo 1000 > serial
 
 
-openssl genrsa -aes256 -out private/ca.key.pem 4096
+openssl genrsa -aes256 -passout file:passfile -out private/ca.key.pem 4096
 chmod 400 private/ca.key.pem
 
 openssl req -config openssl.cnf \
-      -key private/ca.key.pem \
+      -key private/ca.key.pem -passin file:passfile -batch \
       -new -x509 -days 7300 -sha256 -extensions v3_ca \
       -out certs/ca.cert.pem
